requirements.txt
attrs
gevent
